General Cleanup
1. Fixed small typo passing too many args
2. Separated all files into folders
3. Improved Makefile
4. Added .gitignore
