+++ Makefile
@@ -1,21 +1,42 @@
+BIN_DIR = bin
+SRC_DIR = src
+RES_DIR = res
+
 BINARY_PREFIX = LenovoFanControl
-OUT_x64 = ${BINARY_PREFIX}-x64.exe
-OUT_x86 = ${BINARY_PREFIX}-x86.exe
-source_files = lenovo_fan_control.c fanctrl.c fanctrl.h
+OUT_x64 = ${BIN_DIR}/${BINARY_PREFIX}-x64.exe
+OUT_x86 = ${BIN_DIR}/${BINARY_PREFIX}-x86.exe
+
+source_files = ${SRC_DIR}/lenovo_fan_control.c ${SRC_DIR}/fanctrl.c ${SRC_DIR}/fanctrl.h
+res_files = ${RES_DIR}/icon.ico ${RES_DIR}/icon.rc ${RES_DIR}/resource.h
+
+ICON_BIN = ${BIN_DIR}/icon-x64.o
 
 all: ${OUT_x64} ${OUT_x86}
+x64: ${OUT_x64}
+x86: ${OUT_x86}
 
-${OUT_x64}: ${source_files} icon-x64.o
-	gcc -m64 -static -o ${OUT_x64} ${source_files} icon-x64.o -lpthread -mwindows
+# x64
+#########################################################
 
-${OUT_x86}: ${source_files} icon-x86.o
-	gcc -m32 -static -o ${OUT_x86} ${source_files} icon-x86.o -lpthread -mwindows
+${OUT_x64}: ${source_files} ${BIN_DIR}/icon-x64.o
+	gcc -static -m64 -o ${OUT_x64} ${source_files} ${BIN_DIR}/icon-x64.o -lpthread -mwindows
 
-icon-x64.o: icon.ico icon.rc resource.h
-	windres -F pe-x86-64 -i icon.rc -o icon-x64.o
+${BIN_DIR}/icon-x64.o: ${res_files}
+	windres -F pe-x86-64 -i ${RES_DIR}/icon.rc -o ${BIN_DIR}/icon-x64.o
 
-icon-x86.o: icon.ico icon.rc resource.h
-	windres -F pe-i386 -i icon.rc -o icon-x86.o
+# x86
+#########################################################
+
+${OUT_x86}: ${source_files} ${BIN_DIR}/icon-x86.o
+	gcc -static -m32 -o ${OUT_x86} ${source_files} ${BIN_DIR}/icon-x86.o -lpthread -mwindows
+
+${BIN_DIR}/icon-x86.o: ${res_files}
+	windres -F pe-i386 -i ${RES_DIR}/icon.rc -o ${BIN_DIR}/icon-x86.o
+
+#########################################################
+
+clean_all: clean
+	del ${BIN_DIR}\*.exe
 
 clean:
-	del *.o
+	del ${BIN_DIR}\*.o
