+++ README.zh_CN.md
@@ -36,7 +36,7 @@ Language: [English](README.md)
 
 如果你安装了`Lenovo ACPI-Compliant Virtual Power Controller`驱动，那么你的系统中会有一个名为`\\.\EnergyDrv`的设备，此设备正是由这个联想驱动生成的，驱动通过这个设备给其他用户程序提供交互的接口，其他程序可以通过读写这个设备来实现对驱动的控制。这个驱动提供了一个能够控制风扇的除尘功能，我们通过调用win32 API可以轻易地对这个设备的特定字节进行读写，从而控制驱动，让驱动控制EC，最后控制风扇。以下是示意图。
 
-![示意图](images/diagram-zh_CN.jpg)
+![示意图](github/diagram-zh_CN.jpg)
 
 但这样控制风扇有一个问题，就是风扇的运行是断断续续的。当控制驱动执行除尘操作时，风扇会以最高转速运行9秒，然后暂停2秒，然后又运行9秒暂停2秒，循环往复直到2分钟后恢复正常状态。除尘功能是由EC自动控制的，有时它甚至会在风扇高速旋转的9秒中的任意时刻突然停下。
