+++ res/icon.rc
